feat(data_types): add definable decoders & data types for sources
All data sources are normalized to an array of “source” hashes. You can also create source hashes on your own to manipulate their type, alias, & path directly. This way you can take an extensionless path and give it a type.

The decoders used are now configurable, so for example you could override the json or yaml parsers to use OJ for json or safe_yaml for yaml. For any other data type you’re free to define a parser.